Refactor move TreeItemOrderProvider to ProjectSystem.Managed project
