+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/Tree/AppDesignerFolderRenameHandler.cs
@@ -1,12 +1,14 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
+using ManagedPriorityOrder = Microsoft.VisualStudio.ProjectSystem.Order;
+
 namespace Microsoft.VisualStudio.ProjectSystem.Tree
 {
     /// <summary>
     ///     Responsible for setting the AppDesignerFolder property for when the AppDesigner
     ///     folder ("Properties" in C# and "My Project" in Visual Basic) is renamed.
     /// </summary>
-    [Order(Order.Default)]
+    [Order(ManagedPriorityOrder.Default)]
     [Export(typeof(IProjectTreeActionHandler))]
     [AppliesTo(ProjectCapability.AppDesigner)]
     internal class AppDesignerFolderRenameHandler : ProjectTreeActionHandlerBase
