+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/IProjectTreeCustomizablePropertyContextFactory.cs
@@ -8,5 +8,21 @@ namespace Microsoft.VisualStudio.ProjectSystem
         {
             return Mock.Of<IProjectTreeCustomizablePropertyContext>();
         }
+
+        public static IProjectTreeCustomizablePropertyContext Implement(
+            string? itemName = null,
+            string? itemType = null,
+            bool isFolder = false,
+            ProjectTreeFlags flags = default,
+            IImmutableDictionary<string, string>? metadata = null)
+        {
+            var mock = new Mock<IProjectTreeCustomizablePropertyContext>();
+            mock.Setup(x => x.ItemName).Returns(itemName ?? string.Empty);
+            mock.Setup(x => x.ItemType).Returns(itemType);
+            mock.Setup(x => x.IsFolder).Returns(isFolder);
+            mock.Setup(x => x.ParentNodeFlags).Returns(flags);
+            mock.Setup(x => x.Metadata).Returns(metadata ?? ImmutableStringDictionary<string>.EmptyOrdinal);
+            return mock.Object;
+        }
     }
 }
